refactor: tested formatNumber
