+++ src/components/fruitsetlegumes/saisons/list/Item.js
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
-import { formatNumber } from 'utils/formatters'
+import formatNumber from 'utils/formatNumber'
 import { getMonth } from 'utils/months'
 import Emoji from 'components/base/Emoji'
 import MagicLink from 'components/base/MagicLink'
