+++ src/components/misc/tiles/Tile.js
@@ -4,7 +4,8 @@ import React, { useContext } from 'react'
 import CountUp from 'react-countup'
 import styled from 'styled-components'
 import formatName from 'utils/formatName'
-import { formatNumber, formatTotalByMultiplier } from 'utils/formatters'
+import formatNumber from 'utils/formatNumber'
+import { formatTotalByMultiplier } from 'utils/formatters'
 import useIframe from 'hooks/useIframe'
 import DataContext from 'components/providers/DataProvider'
 import Button from 'components/base/Button'
