+++ src/components/numerique/category/result/Detail.js
@@ -1,7 +1,8 @@
 import { useContext, useMemo, useState } from 'react'
 import styled from 'styled-components'
 import formatName from 'utils/formatName'
-import { formatConstruction, formatNumber } from 'utils/formatters'
+import formatNumber from 'utils/formatNumber'
+import { formatConstruction } from 'utils/formatters'
 import DataContext from 'components/providers/DataProvider'
 import Checkbox from 'components/base/Checkbox'
 import BarChart from 'components/charts/BarChart'
