+++ src/components/numerique/equivalent/StockageEmails.js
@@ -1,6 +1,7 @@
 import React, { useContext, useMemo, useState } from 'react'
 import styled from 'styled-components'
-import { formatNumber, formatTotal } from 'utils/formatters'
+import formatNumber from 'utils/formatNumber'
+import { formatTotal } from 'utils/formatters'
 import DataContext from 'components/providers/DataProvider'
 import Button from 'components/base/Button'
 import ButtonLink from 'components/base/ButtonLink'
