+++ src/components/views/equivalent/details/Value.js
@@ -1,6 +1,7 @@
 import React from 'react'
 import styled from 'styled-components'
-import { formatNumber, formatTotal } from 'utils/formatters'
+import formatNumber from 'utils/formatNumber'
+import { formatTotal } from 'utils/formatters'
 import getFrenchFormattedNumber from 'utils/getFrenchFormattedNumber'
 import getNumberWithNDigitsAfterComma from 'utils/getNumberWithNDigitsAfterComma'
 import Emoji from 'components/base/Emoji'
